requirements.txt
boto3==1.16.10
botocore==1.19.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pyasn1==0.4.8
python-dateutil==2.8.1
python-jose==3.2.0
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
